+++ shuup/core/suppliers/enums.py
@@ -12,7 +12,9 @@ from enumfields import Enum
 class StockAdjustmentType(Enum):
     INVENTORY = 1
     RESTOCK = 2
+    RESTOCK_LOGICAL = 3
 
     class Labels:
         INVENTORY = _("inventory")
         RESTOCK = _("restock")
+        RESTOCK_LOGICAL = _("restock logical")