+++ shuup/simple_supplier/utils.py
@@ -14,6 +14,7 @@ from django.template.loader import render_to_string
 from shuup.core.models import (
     OrderLine, OrderLineType, OrderStatusRole, ShipmentProduct
 )
+from shuup.core.suppliers.base import StockAdjustmentType
 from shuup.simple_supplier.forms import StockAdjustmentForm
 from shuup.simple_supplier.models import StockAdjustment, StockCount
 
@@ -36,24 +37,27 @@ def get_current_stock_value(supplier_id, product_id):
     events = (
         StockAdjustment.objects
         .filter(supplier_id=supplier_id, product_id=product_id)
+        .exclude(type=StockAdjustmentType.RESTOCK_LOGICAL)
         .aggregate(total=Sum("delta"))["total"] or 0)
-    orders_bought = (
+    products_bought = (
         OrderLine.objects
         .filter(supplier_id=supplier_id, product_id=product_id)
         .exclude(
             Q(order__status__role=OrderStatusRole.CANCELED) |
-            Q(type=OrderLineType.AMOUNT_REFUND) |
-            Q(type=OrderLineType.QUANTITY_REFUND)
-        )
+            Q(type=OrderLineType.REFUND))
         .aggregate(total=Sum("quantity"))["total"] or 0)
-    orders_sent = (
+    products_refunded_before_shipment = (
+        StockAdjustment.objects
+        .filter(supplier_id=supplier_id, product_id=product_id, type=StockAdjustmentType.RESTOCK_LOGICAL)
+        .aggregate(total=Sum("delta"))["total"] or 0)
+    products_sent = (
         ShipmentProduct.objects
         .filter(shipment__supplier=supplier_id, product_id=product_id)
         .aggregate(total=Sum("quantity"))["total"] or 0)
 
     return {
-        "logical_count": events - orders_bought,
-        "physical_count": events - orders_sent
+        "logical_count": events - products_bought + products_refunded_before_shipment,
+        "physical_count": events - products_sent
     }
 
 
@@ -98,7 +102,7 @@ def get_stock_adjustment_div(request, supplier, product):
     :rtype: str
     """
     latest_adjustment = StockAdjustment.objects.filter(
-        product=product, supplier=supplier).order_by("-created_on").first()
+        product=product, supplier=supplier, type=StockAdjustmentType.INVENTORY).last()
     purchase_price = (latest_adjustment.purchase_price_value if latest_adjustment else Decimal("0.00"))
     context = {
         "product": product,