+++ shuup/testing/factories.py
@@ -531,11 +531,14 @@ def add_product_to_order(order, supplier, product, quantity, taxless_base_unit_p
                                    order_line=product_order_line,
                                    product=product, quantity=quantity,
                                    supplier=supplier)
-    product_order_line.base_unit_price = order.shop.create_price(taxless_base_unit_price)
+    base_unit_price = order.shop.create_price(taxless_base_unit_price)
+    if order.prices_include_tax:
+        base_unit_price *= (1 + tax_rate)
+    product_order_line.base_unit_price = order.shop.create_price(base_unit_price)
     product_order_line.save()
     product_order_line.taxes.add(OrderLineTax.from_tax(
         get_test_tax(tax_rate),
-        product_order_line.taxless_price.amount,
+        Money(quantity * taxless_base_unit_price, order.currency),
         order_line=product_order_line,
     ))