+++ shuup_tests/core/test_orders.py
@@ -14,18 +14,18 @@ from django.test import override_settings
 from django.utils.timezone import now
 
 from shuup.core.excs import (
-    NoPaymentToCreateException, NoProductsToShipException,
-    RefundExceedsAmountException
+    InvalidRefundAmountException, NoPaymentToCreateException,
+    NoProductsToShipException, RefundExceedsAmountException
 )
 from shuup.core.models import (
-    Order, OrderLine, OrderLineTax, OrderLineType, OrderStatus,
-    PaymentStatus, ShippingStatus, StockBehavior
+    Order, OrderLine, OrderLineTax, OrderLineType, OrderStatus, PaymentStatus,
+    ShippingStatus, StockBehavior
 )
 from shuup.core.pricing import TaxfulPrice, TaxlessPrice
 from shuup.testing.factories import (
     create_empty_order, create_order_with_product, create_product, get_address,
     get_default_product, get_default_shop, get_default_supplier,
-    get_default_tax, get_initial_order_status, get_shipping_method
+    get_default_tax, get_initial_order_status
 )
 from shuup.utils.money import Money
 from shuup_tests.simple_supplier.utils import get_simple_supplier
@@ -253,6 +253,7 @@ def test_payments():
     assert order.payment_status == PaymentStatus.FULLY_PAID
     assert not order.can_edit()
 
+
 @pytest.mark.django_db
 def test_refunds():
     shop = get_default_shop()
@@ -263,7 +264,8 @@ def test_refunds():
         default_price=10,
     )
     tax_rate = Decimal("0.1")
-    order = create_order_with_product(product, supplier, 2, 200, tax_rate, shop=shop)
+    taxless_base_unit_price = shop.create_price(200)
+    order = create_order_with_product(product, supplier, 3, taxless_base_unit_price, tax_rate, shop=shop)
     order.payment_status = PaymentStatus.DEFERRED
     order.cache_prices()
     order.save()
@@ -279,46 +281,42 @@ def test_refunds():
     assert order.can_create_refund()
     assert not order.has_refunds()
 
-    # Create a refund with a parent line and quantity
-    order.create_refund([{"line": product_line, "quantity": 1}])
+    order.create_refund([{"line": product_line, "quantity": 1, "amount": (product_line.taxful_price.amount / 3)}])
     assert len(order.lines.all()) == 2
     assert order.lines.last().ordering == 1
     assert order.has_refunds()
 
     # Confirm the value of the refund
     assert order.lines.last().taxful_price == -product_line.base_unit_price
-    assert order.lines.last().tax_amount == -(product_line.base_unit_price * tax_rate).amount
-
-    partial_refund_amount = order.taxful_total_price.amount / 2
-    remaining_amount = order.taxful_total_price.amount - partial_refund_amount
+    assert order.lines.last().tax_amount == -(product_line.taxless_base_unit_price * tax_rate).amount
 
     # Create a refund with a parent line and amount
-    order.create_refund([{"line": product_line, "amount": partial_refund_amount}])
+    order.create_refund([{"line": product_line, "quantity": 1, "amount": product_line.taxful_price.amount / 3}])
     assert len(order.lines.all()) == 3
     assert order.lines.last().ordering == 2
 
-    assert order.lines.last().taxful_price.amount == -partial_refund_amount
-    assert order.lines.last().tax_amount == -partial_refund_amount * tax_rate
+    assert order.lines.last().taxful_price.amount == -taxless_base_unit_price.amount * (1 + tax_rate)
+    assert order.lines.last().tax_amount == -taxless_base_unit_price.amount * tax_rate
 
-    assert order.taxless_total_price.amount == remaining_amount * (1 - tax_rate)
-    assert order.taxful_total_price.amount == remaining_amount
+    assert order.taxless_total_price.amount == taxless_base_unit_price.amount
+    assert order.taxful_total_price.amount == taxless_base_unit_price.amount * (1 + tax_rate)
     assert order.can_create_refund()
 
     # Try to refunding remaining amount without a parent line
     with pytest.raises(AssertionError):
-        order.create_refund([{"amount": remaining_amount}])
+        order.create_refund([{"amount": taxless_base_unit_price}])
 
     # refund remaining amount
-    order.create_refund([{"line": product_line, "amount": remaining_amount}])
+    order.create_refund([{"line": product_line, "quantity": 1, "amount": product_line.taxful_price.amount / 3}])
     assert len(order.lines.all()) == 4
     assert order.lines.last().ordering == 3
-    assert order.lines.last().taxful_price.amount == -remaining_amount
+    assert order.lines.last().taxful_price.amount == -taxless_base_unit_price.amount * (1 + tax_rate)
 
     assert not order.taxful_total_price.amount
     assert not order.can_create_refund()
 
     with pytest.raises(RefundExceedsAmountException):
-        order.create_refund([{"line": product_line, "amount": remaining_amount}])
+        order.create_refund([{"line": product_line, "quantity": 1, "amount": taxless_base_unit_price.amount}])
 
 
 @pytest.mark.django_db
@@ -334,8 +332,9 @@ def test_refund_entire_order():
     supplier.adjust_stock(product.id, 5)
     check_stock_counts(supplier, product, 5, 5)
 
-    order = create_order_with_product(product, supplier, 2, 200, Decimal("0.1"), shop=shop)
+    order = create_order_with_product(product, supplier, 2, 200, Decimal("0.24"), shop=shop)
     order.cache_prices()
+
     original_total_price = order.taxful_total_price
     check_stock_counts(supplier, product, 5, 3)
 
@@ -346,7 +345,7 @@ def test_refund_entire_order():
     assert order.taxless_total_price.amount.value == 0
     assert order.taxful_total_price.amount.value == 0
     refund_line = order.lines.order_by("ordering").last()
-    assert refund_line.type == OrderLineType.QUANTITY_REFUND
+    assert refund_line.type == OrderLineType.REFUND
     assert refund_line.taxful_price == -original_total_price
 
     # Make sure logical count reflects refunded products
@@ -368,16 +367,14 @@ def test_refund_entire_order_with_product_restock():
 
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
     order.cache_prices()
-    original_total_price = order.taxful_total_price
 
     check_stock_counts(supplier, product, 5, 3)
 
     # Create a full refund with `restock_products` set to True
     order.create_full_refund(restock_products=True)
 
-    # Since no shipments were created, restocking is not possible
-    check_stock_counts(supplier, product, 5, 3)
-
+    # restock logical count
+    check_stock_counts(supplier, product, 5, 5)
 
 
 @pytest.mark.django_db
@@ -404,27 +401,29 @@ def test_refund_with_shipment(restock):
     # Shipment should decrease physical count by 2, logical by none
     order.create_shipment({product_line.product: 2}, supplier=supplier)
     check_stock_counts(supplier, product, physical=8, logical=6)
-    assert order.shipping_status ==  ShippingStatus.PARTIALLY_SHIPPED
+    assert order.shipping_status == ShippingStatus.PARTIALLY_SHIPPED
 
     # Check correct refunded quantities
     assert not product_line.refunded_quantity
 
-    # Create a refund greater than restockable quantity, check stocks
+    # Create a refund that refunds from unshipped quantity first, then shipped quantity, check stocks
     check_stock_counts(supplier, product, physical=8, logical=6)
-    order.create_refund([{"line": product_line, "quantity": 3, "restock_products": restock}])
+    order.create_refund([
+        {"line": product_line, "quantity": 3, "amount": Money(600, order.currency), "restock_products": restock}])
     assert product_line.refunded_quantity == 3
     assert order.shipping_status == ShippingStatus.FULLY_SHIPPED
     if restock:
-        check_stock_counts(supplier, product, physical=10, logical=8)
+        check_stock_counts(supplier, product, physical=9, logical=9)
     else:
         check_stock_counts(supplier, product, physical=8, logical=6)
 
-    # Create a second refund greater than restockable quantity, check stocks
-    order.create_refund([{"line": product_line, "quantity": 1, "restock_products": restock}])
+    # Create a second refund that refunds the last shipped quantity, check stocks
+    order.create_refund([
+        {"line": product_line, "quantity": 1, "amount": Money(200, order.currency), "restock_products": restock}])
     assert product_line.refunded_quantity == 4
     if restock:
         # Make sure we're not restocking more than maximum restockable quantity
-        check_stock_counts(supplier, product, physical=10, logical=8)
+        check_stock_counts(supplier, product, physical=10, logical=10)
     else:
         # Make sure maximum restockable quantity is not 0
         check_stock_counts(supplier, product, physical=8, logical=6)
@@ -451,10 +450,13 @@ def test_refund_without_shipment(restock):
 
     # Restock value shouldn't matter if we don't have any shipments
     product_line = order.lines.first()
-    order.create_refund([{"line": product_line, "quantity": 2, "restock_products": restock}])
+    order.create_refund([
+        {"line": product_line, "quantity": 2, "amount": Money(400, order.currency), "restock_products": restock}])
 
-    check_stock_counts(supplier, product, physical=10, logical=8)
-
+    if restock:
+        check_stock_counts(supplier, product, physical=10, logical=10)
+    else:
+        check_stock_counts(supplier, product, physical=10, logical=8)
     assert product_line.refunded_quantity == 2
 
 
@@ -468,7 +470,7 @@ def test_max_refundable_amount():
         default_price=10,
     )
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
-
+    order.cache_prices()
     assert len(order.lines.all()) == 1
 
     line = order.lines.first()
@@ -477,7 +479,7 @@ def test_max_refundable_amount():
     partial_refund_amount = Money(10, order.currency)
     assert partial_refund_amount.value > 0
 
-    order.create_refund([{"line": line, "amount": partial_refund_amount}])
+    order.create_refund([{"line": line, "quantity": 1, "amount": partial_refund_amount}])
     assert line.max_refundable_amount == line.taxful_price.amount - partial_refund_amount
 
 
@@ -493,45 +495,50 @@ def test_refunds_for_discounted_order_lines():
     )
 
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
-    order.cache_prices()
+    discount_line = OrderLine(type=OrderLineType.DISCOUNT, quantity=1, discount_amount_value=Decimal("0.54321"))
+    order.lines.add(discount_line)
 
-    product_line = order.lines.first()
+    product_line = order.lines.filter(type=OrderLineType.PRODUCT).first()
     product_line.discount_amount = TaxfulPrice(100, order.currency)
+    product_line.save()
     taxful_price_with_discount = product_line.taxful_price
+    order.cache_prices()
+    order.save()
+
     assert product_line.base_price == TaxfulPrice(400, order.currency)
     assert taxful_price_with_discount == TaxfulPrice(300, order.currency)
 
-    order.create_refund([{"line": product_line, "quantity": 1}])
-    assert order.lines.refunds().first().taxful_price == (-taxful_price_with_discount / 2)
+    # try to refund only the product line - should fail since this would result in a negative total
+    with pytest.raises(RefundExceedsAmountException):
+        order.create_refund([{"line": product_line, "quantity": 2, "amount": taxful_price_with_discount.amount}])
 
+    # try to refund the product line with a negative amount
+    with pytest.raises(InvalidRefundAmountException):
+        order.create_refund([{"line": product_line, "quantity": 1, "amount": -taxful_price_with_discount.amount}])
+    # try to refund the discount line with a positive amount
+    with pytest.raises(InvalidRefundAmountException):
+        order.create_refund([{"line": discount_line, "quantity": 1, "amount": -discount_line.taxful_price.amount}])
 
-@pytest.mark.django_db
-def test_refunds_with_quantities():
-    shop = get_default_shop()
-    supplier = get_default_supplier()
-    product = create_product(
-        "test-sku",
-        shop=get_default_shop(),
-        default_price=10,
-        stock_behavior=StockBehavior.STOCKED
-    )
+    order.create_refund([
+        {"line": discount_line, "quantity": 1, "amount": discount_line.taxful_price.amount},
+        {"line": product_line, "quantity": 2, "amount": taxful_price_with_discount.amount}
+    ])
+    assert product_line.max_refundable_amount.value == 0
+    assert discount_line.max_refundable_amount.value == 0
+    assert order.taxful_total_price.value == 0
 
-    order = create_order_with_product(product, supplier, 3, 200, shop=shop)
+    order = create_order_with_product(product, supplier, 2, 200, shop=shop)
+    discount_line = OrderLine(type=OrderLineType.DISCOUNT, quantity=1, discount_amount_value=Decimal("0.54321"))
+    order.lines.add(discount_line)
+    product_line = order.lines.filter(type=OrderLineType.PRODUCT).first()
+    product_line.discount_amount = TaxfulPrice(100, order.currency)
+    product_line.save()
     order.cache_prices()
-    assert not order.lines.refunds()
+    order.save()
 
-    product_line = order.lines.first()
-    refund_amount = Money(100, order.currency)
-    order.create_refund([{"line": product_line, "quantity": 2, "amount": refund_amount}])
-    assert order.lines.refunds().count() == 2
+    order.create_full_refund(restock_products=False)
+    assert order.taxful_total_price.value == 0
 
-    quantity_line = order.lines.refunds().filter(quantity=2).first()
-    assert quantity_line
-    amount_line = order.lines.refunds().filter(quantity=1).first()
-    assert amount_line
-
-    assert quantity_line.taxful_base_unit_price == -product_line.taxful_base_unit_price
-    assert amount_line.taxful_price.amount == -refund_amount
 
 @pytest.mark.django_db
 def test_can_create_shipment():
@@ -585,6 +592,7 @@ def test_can_create_payment():
     assert not order.can_create_payment()
 
     order = create_order_with_product(product, supplier, 1, 200, shop=shop)
+    order.cache_prices()
     assert order.can_create_payment()
 
     # Canceled orders can't create payments
@@ -592,14 +600,17 @@ def test_can_create_payment():
     assert not order.can_create_payment()
 
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
+    order.cache_prices()
     assert order.can_create_payment()
 
     # Partially refunded orders can create payments
-    order.create_refund([{"line": order.lines.first(), "quantity": 1, "restock": False}])
+    order.create_refund([
+        {"line": order.lines.first(), "quantity": 1, "amount": Money(200, order.currency), "restock": False}])
     assert order.can_create_payment()
 
     # But fully refunded orders can't
-    order.create_refund([{"line": order.lines.first(), "quantity": 1, "restock": False}])
+    order.create_refund([
+        {"line": order.lines.first(), "quantity": 1, "amount": Money(200, order.currency), "restock": False}])
     assert not order.can_create_payment()
 
 
@@ -615,14 +626,17 @@ def test_can_create_refund():
     )
 
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
+    order.cache_prices()
     assert order.can_create_payment()
 
     # Partially refunded orders can create refunds
-    order.create_refund([{"line": order.lines.first(), "quantity": 1, "restock": False}])
+    order.create_refund([
+        {"line": order.lines.first(), "quantity": 1, "amount": Money(200, order.currency), "restock": False}])
     assert order.can_create_refund()
 
     # But fully refunded orders can't
-    order.create_refund([{"line": order.lines.first(), "quantity": 1, "restock": False}])
+    order.create_refund([
+        {"line": order.lines.first(), "quantity": 1, "amount": Money(200, order.currency), "restock": False}])
     assert not order.can_create_refund()
 
 
@@ -644,9 +658,9 @@ def test_product_summary():
 
     # Order with 2 unshipped, non-refunded items and a shipping cost
     order = create_order_with_product(product, supplier, 2, 200, shop=shop)
+    order.cache_prices()
     product_line = order.lines.first()
-    sm = get_shipping_method(name="test", price=10)
-    shipping_line  = order.lines.create(type=OrderLineType.SHIPPING, base_unit_price_value=5, quantity=1)
+    shipping_line = order.lines.create(type=OrderLineType.SHIPPING, base_unit_price_value=5, quantity=1)
 
     # Make sure no invalid entries and check product quantities
     product_summary = order.get_product_summary()
@@ -660,7 +674,7 @@ def test_product_summary():
     order.create_shipment(supplier=supplier, product_quantities={product: 1})
     assert order.shipping_status == ShippingStatus.PARTIALLY_SHIPPED
 
-    order.create_refund([{"line": shipping_line, "quantity": 1, "restock": False}])
+    order.create_refund([{"line": shipping_line, "quantity": 1, "amount": Money(5, order.currency), "restock": False}])
 
     product_summary = order.get_product_summary()
     assert all(product_summary.keys())
@@ -668,7 +682,7 @@ def test_product_summary():
     assert_defaultdict_values(summary, ordered=2, shipped=1, refunded=0, unshipped=1)
 
     # Create a refund for 2 items, we should get no negative values
-    order.create_refund([{"line": product_line, "quantity": 2, "restock": False}])
+    order.create_refund([{"line": product_line, "quantity": 2, "amount": Money(200, order.currency), "restock": False}])
 
     product_summary = order.get_product_summary()
     assert all(product_summary.keys())